Fixing issues and adding tests

+++ src/test/java/com/emerchantpay/paymentsystemtask/dao/TransactionRepositoryTest.java
@@ -8,6 +8,8 @@ import com.emerchantpay.paymentsystemtask.model.transaction.AuthorizeTransaction
 import com.emerchantpay.paymentsystemtask.model.transaction.ChargeTransaction;
 import com.emerchantpay.paymentsystemtask.model.transaction.Transaction;
 import com.emerchantpay.paymentsystemtask.utils.TransactionUtils;
+import org.apache.juli.logging.Log;
+import org.apache.juli.logging.LogFactory;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,11 +22,11 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-
 @RunWith(SpringRunner.class)
 @SpringBootTest
 public class TransactionRepositoryTest {
 
+    private final Log log = LogFactory.getLog(this.getClass());
     @Autowired
     private TransactionRepository transRepository;
 
@@ -36,24 +38,26 @@ public class TransactionRepositoryTest {
     public void findAuthorizeTransactionByRefIdMerId_success(){
 
         Merchant mer = PaymentsUtils.createMerchant();
-        merchantRepository.save(mer);
+        Transaction authTrans = PaymentsUtils.setTransactionProperties(TransactionType.AUTHORIZE,null, mer);
+        Transaction chTrans =  PaymentsUtils.setTransactionProperties(TransactionType.CHARGE, authTrans.getUuid(), authTrans.getMerchant());
+        try {
+            merchantRepository.save(mer);
+            transRepository.saveAll(List.of(authTrans, chTrans));
 
-        PaymentsUtils paymentsUtils=new PaymentsUtils();
-        Transaction authTrans =
-                paymentsUtils.createTransaction(TransactionType.AUTHORIZE.getName(),null,mer);
-        Transaction chargeTransaction =
-                paymentsUtils.createTransaction(TransactionType.CHARGE.getName(), authTrans.getUuid(),mer);
-        List<Transaction> trans =List.of(authTrans,chargeTransaction);
-        transRepository.saveAll(trans);
-
-        AuthorizeTransaction auth =
-                transRepository.findAuthorizeTransactionByRefIdMerId(chargeTransaction.getReferenceIdentifier(),
-                        chargeTransaction.getStatus(),
-                        String.valueOf(chargeTransaction.getMerchant().getIdentifier()));
-        assertNotNull(auth);
-
-        transRepository.deleteAllById(List.of(chargeTransaction.getUuid(),authTrans.getUuid()));
-        merchantRepository.deleteById(mer.getIdentifier());
+            AuthorizeTransaction auth =
+                    transRepository.findAuthorizeTransactionByRefIdMerId(chTrans.getReferenceIdentifier(),
+                            chTrans.getStatus(),
+                            String.valueOf(chTrans.getMerchant().getIdentifier()));
+            assertNotNull(auth);
+        } catch (Exception e){
+            log.error(e.getMessage());
+            fail(e.getMessage());
+        } finally {
+            transRepository.deleteAllById(List.of(chTrans.getUuid(),authTrans.getUuid()));
+            if(mer.getIdentifier()!=null){
+                merchantRepository.deleteById(mer.getIdentifier());
+            }
+        }
     }
 
     @Test
@@ -70,27 +74,30 @@ public class TransactionRepositoryTest {
     public void findChargeTransactionByRefIdMerIdTest_success(){
 
         Merchant mer = PaymentsUtils.createMerchant();
-        merchantRepository.save(mer);
+        Transaction authTrans =  PaymentsUtils.setTransactionProperties(TransactionType.AUTHORIZE,null,mer);
+        Transaction chTrans =  PaymentsUtils.setTransactionProperties(TransactionType.CHARGE,authTrans.getUuid(),authTrans.getMerchant());
+        Transaction refTrans = PaymentsUtils.setTransactionProperties(TransactionType.REFUND,chTrans.getReferenceIdentifier(),chTrans.getMerchant());
 
-        PaymentsUtils paymentsUtils=new PaymentsUtils();
-        Transaction authTrans =
-                paymentsUtils.createTransaction(TransactionType.AUTHORIZE.getName(),null,mer);
-        Transaction chargeTransaction =
-                paymentsUtils.createTransaction(TransactionType.CHARGE.getName(), authTrans.getUuid(),authTrans.getMerchant());
-        Transaction refundTrans=
-                paymentsUtils.createTransaction(TransactionType.REFUND.getName(), chargeTransaction.getReferenceIdentifier(),chargeTransaction.getMerchant());
-        List<Transaction> trans =List.of(authTrans,chargeTransaction,refundTrans);
-        transRepository.saveAll(trans);
+        List<Transaction> trans = List.of(authTrans,chTrans,refTrans);
+        try {
+            merchantRepository.save(mer);
+            transRepository.saveAll(trans);
 
-        ChargeTransaction ch =
-                transRepository.findChargeTransactionByRefIdMerId(refundTrans.getReferenceIdentifier(),
-                        refundTrans.getStatus(),
-                        String.valueOf(refundTrans.getMerchant().getIdentifier()));
-        assertNotNull(ch);
+            ChargeTransaction ch =
+                    transRepository.findChargeTransactionByRefIdMerId(refTrans.getReferenceIdentifier(),
+                            refTrans.getStatus(),
+                            String.valueOf(refTrans.getMerchant().getIdentifier()));
+            assertNotNull(ch);
 
-        transRepository.deleteAllById(List.of(refundTrans.getUuid(),chargeTransaction.getUuid(),authTrans.getUuid()));
-        merchantRepository.deleteById(mer.getIdentifier());
-
+        } catch (Exception e){
+            log.error(e.getMessage());
+            fail(e.getMessage());
+        } finally {
+            transRepository.deleteAllById(List.of(refTrans.getUuid(),chTrans.getUuid(),authTrans.getUuid()));
+            if(mer.getIdentifier()!=null){
+                merchantRepository.deleteById(mer.getIdentifier());
+            }
+        }
     }
 
     @Test
@@ -106,19 +113,16 @@ public class TransactionRepositoryTest {
 
     @Test
     public void existsByTransactionTypeAndReferenceIdentifier_success(){
-        PaymentsUtils paymentsUtils=new PaymentsUtils();
-        Transaction authTrans =
-                paymentsUtils.createTransaction(TransactionType.AUTHORIZE.getName(),null,null);
 
-        Transaction chargeTransaction =
-                paymentsUtils.createTransaction(TransactionType.CHARGE.getName(), authTrans.getUuid(),authTrans.getMerchant());
+        Transaction authTrans = PaymentsUtils.setTransactionProperties(TransactionType.AUTHORIZE,null,null);
+        Transaction chTrans = PaymentsUtils.setTransactionProperties(TransactionType.CHARGE,authTrans.getUuid(),null);
+        transRepository.saveAll(List.of(authTrans,chTrans));
 
-        transRepository.saveAll(List.of(authTrans,chargeTransaction));
         boolean isTransExist
-                = transRepository.existsByTransactionTypeAndReferenceIdentifier(chargeTransaction.getTransactionType(),
-                                                                                 chargeTransaction.getReferenceIdentifier());
+                = transRepository.existsByTransactionTypeAndReferenceIdentifier(chTrans.getTransactionType(),
+                                                                                 chTrans.getReferenceIdentifier());
         assertTrue(isTransExist);
-        transRepository.deleteAllById(List.of(chargeTransaction.getUuid(),authTrans.getUuid()));
+        transRepository.deleteAllById(List.of(chTrans.getUuid(),authTrans.getUuid()));
     }
 
 
@@ -135,10 +139,7 @@ public class TransactionRepositoryTest {
     @Test
     public void deleteTranOlderThanHourTest_success(){
 
-        PaymentsUtils paymentsUtils=new PaymentsUtils();
-        Transaction authTrans =
-                paymentsUtils.createTransaction(TransactionType.AUTHORIZE.getName(),null,null);
-
+        Transaction authTrans = PaymentsUtils.setTransactionProperties(TransactionType.AUTHORIZE,null, null);
         Calendar calendar = Calendar.getInstance();
         calendar.add(Calendar.HOUR, -2);
         Timestamp time=authTrans.getTimestamp();
@@ -153,9 +154,7 @@ public class TransactionRepositoryTest {
     @Test
     public void deleteTranOlderThanHourTest_fail() {
 
-        PaymentsUtils paymentsUtils=new PaymentsUtils();
-        Transaction authTrans =
-                paymentsUtils.createTransaction(TransactionType.AUTHORIZE.getName(), null, null);
+        Transaction authTrans = PaymentsUtils.setTransactionProperties(TransactionType.AUTHORIZE,null,null);
 
         transRepository.save(authTrans);
